rectified quicksort bug

+++ algorithm/sort/includes/QuickSort.hpp
@@ -3,8 +3,8 @@
 #include <cstddef>
 #include <vector>
 
-void QuickSort(std::vector <size_t> & arr, const size_t & low, const size_t & high);
-void QuickSortIterative(std::vector <size_t> & arr, const size_t & low, const size_t & high);
+void QuickSort(std::vector <size_t> & arr, const int64_t & low, const int64_t & high);
+void QuickSortIterative(std::vector <size_t> & arr, const int64_t & low, const int64_t & high);
 
 #endif /* QUICKSORT_HPP */
